layout/generic/test/test_bug831780.html


<html>



<head>
  <meta charset="utf-8">
  <title>Test for Bug 831780</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=831780">Mozilla Bug 831780</a>
<p id="display" style="width: 1px; height: 1px; overflow: hidden">
  

  <img style="width: 1px; height: 1px">
</p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
  <script type="application/javascript">

  /** Test for Bug 831780 **/
  is($("display").scrollWidth, 1, "scrollWidth should be 1");
  is($("display").scrollHeight, 1, "scrollHeight should be 1");

  </script>
</body>
</html>
